feat: move OVERRIDE -> BREK for clarity and to not pollute ENV namespace...
Reasoning is that the word 'override' was being overused. Not only does it mean
a CLI/ENV value, but it also can mean any non-default setting in users,
environment, or deployment. All of the above is an 'override' in brek's
parliance, so we're rennaming OVERRIDE to BREK. OVERRIDE will still work
as this is a non-breaking change, but it is now deprecated.